+++ OopsAllArrowsMod/OopsAllArrowsModSettings.cs
@@ -2,7 +2,7 @@ using Monocle;
 using FortRise;
 using TowerFall;
 namespace OopsAllArrowsMod;
-public class ExampleModSettings : ModuleSettings 
+public class OopsAllArrowsModSettings : ModuleSettings 
 {
 
 }
